feat ✨: Add requirements for ROCm in rocm.txt
This commit introduces a new requirements file, `requirements-rocm.txt`, which specifies the Python package `onnxruntime_rocm-1.17.0-cp310-cp310-linux_x86_64.whl` as a dependency for ROCm environments. This ensures that users with ROCm support can install and use this specific version of OnnxRuntime, which is compatible with their hardware architecture.